requirements.txt
beautifulsoup4
msgpack_python
packaging
requests
SQLAlchemy
python-telegram-bot
loguru
